+++ np.MeetingNotes/src/NPMeetingNotes.js
@@ -53,7 +53,7 @@ export async function insertNoteTemplate(origFileName: string, dailyNoteDate: Da
     const note = DataStore.calendarNoteByDate(dailyNoteDate, timeframe)
     if (note) {
       if (note.content && note.content !== '') {
-        note.content += '\n\n' + result
+        note.content += `\n\n${result}`
       } else {
         note.content = result
       }
@@ -261,6 +261,9 @@ async function handleExistingNotes(_noteTitle: string, renderedContent: string, 
     }
   } else {
     logDebug(pluginJson, `handleExistingNotes: creating note with content:"${noteContent}"`)
+    if (/choose|select/i.test(folder)) {
+      folder = await chooseFolder('Choose a folder to create note in', false, true)
+    }
     noteTitle = (await newNoteWithFolder(noteContent, folder)) ?? '<error>'
   }
   return noteTitle
@@ -354,11 +357,15 @@ function writeNoteLinkIntoEvent(selectedEvent: TCalendarItem, newTitle: string):
  * @param {string} folder - The folder where the note is located.
  * @returns {Promise<CoreNoteFields>} The note.
  */
-async function getNoteBasedOnName(noteName: string, folder: string): Promise<CoreNoteFields> {
+async function getNoteBasedOnName(noteName: string, folder: string): Promise<CoreNoteFields | null> {
+  logDebug(`np.MeetingNotes getNoteBasedOnName: "${noteName}" folder: ${folder}`)
   if (noteName === '<select>') {
     return await getNoteFromSelection(folder)
-  } else if (noteName === '<current>') {
+  } else if (/<current>/i.test(noteName)) {
     return getNoteFromEditor()
+  } else if (/<today>/i.test(noteName)) {
+    await Editor.openNoteByDate(new Date())
+    return Editor
   } else {
     return getNoteByTitle(noteName, folder)
   }
@@ -404,10 +411,10 @@ function getNoteFromEditor(): CoreNoteFields {
  * @param {string} folder - The folder where the note is located.
  * @returns {CoreNoteFields} The note.
  */
-function getNoteByTitle(noteName: string, folder: string): CoreNoteFields {
+function getNoteByTitle(noteName: string, folder: string): CoreNoteFields | null {
   const availableNotes = DataStore.projectNoteByTitle(noteName)
   if (availableNotes && availableNotes.length > 0) {
-    if (folder) {
+    if (folder && !/choose|select/i.test(folder)) {
       const filteredNotes = availableNotes?.filter((n) => n.filename.startsWith(folder)) ?? []
       if (filteredNotes.length > 0) {
         return filteredNotes[0]
@@ -474,10 +481,13 @@ async function updateNoteContent(note: CoreNoteFields, location: string, content
  * @param {string} content - The new content.
  * @returns {Promise<string|null>} The title of the note or null.
  */
-async function appendPrependNewNote(noteName: string, location: string, folder: string = '', content: string): Promise<string | null> {
+async function appendPrependNewNote(noteName: string, location: string, _folder: string = '', content: string): Promise<string | null> {
   try {
+    let folder = _folder
+    logDebug(`np.MeetingNotes appendPrependNewNote noteName=${noteName} location:${location} folder:${folder}`)
     let note = await getNoteBasedOnName(noteName, folder)
     if (!note) {
+      if (/<choose>|<select>/i.test(folder)) folder = await chooseFolder('Choose folder to create note in', false, true)
       note = await createNewNoteIfNotFound(noteName, folder)
     }